Adds console access through a rake task
Sometimes it is nice when working on a gem to be able to start an irb
session with everything already loaded. Now you don't have to do a
require as soon as you start. It also doesn't need to have the gem
installed to work.

Amos King @adkron <amos@binarynoggin.com>